update
Former-commit-id: 13760e265f320f09a9de47a49c32a73c6d3c1be7

+++ trip/grid/gridio.c
@@ -543,12 +543,12 @@ int rsfread(ireal * a,
     for (i=0;i<noffs;i++) {
       /* seek to read segment */
       if (!err && fseeko(fp,goffs[i]*sizeof(float) + cur_pos,SEEK_SET)) {
-	fprintf(stream,"Error: rsfread from fseeko at file offset %lld\n",(intmax_t)goffs[i]); 
+	fprintf(stream,"Error: rsfread from fseeko at file offset %lld\n",(long long int)goffs[i]); 
 	err=E_FILE;
       }
       /* read in byte string */
       if (!err && (gl_na[0] != (int) fread(fbuf,sizeof(float),gl_na[0],fp))) {
-	fprintf(stream,"Error: rsfread from fread at array offset %lld\n",(intmax_t)loffs[i]);
+	fprintf(stream,"Error: rsfread from fread at array offset %lld\n",(long long int)loffs[i]);
 	err=E_FILE;
       }
       /* convert to ireal */
@@ -580,12 +580,12 @@ int rsfread(ireal * a,
 
       /* seek to read segment */
       if (!err && fseeko(fp,goffs[i]*sizeof(float) + cur_pos,SEEK_SET)) {
-	fprintf(stream,"Error: rsfread from fseeko at file offset %lld\n",(intmax_t)goffs[i]);
+	fprintf(stream,"Error: rsfread from fseeko at file offset %lld\n",(long long int)goffs[i]);
 	err=E_FILE;
       }
       /* read in byte array */
       if (!err && (recsize_b != fread(buf,sizeof(char),recsize_b,fp))) {
-	fprintf(stream,"Error: rsfread from fread at array offset %lld\n",(intmax_t)loffs[i]);
+	fprintf(stream,"Error: rsfread from fread at array offset %lld\n",(long long int)loffs[i]);
 	err=E_FILE;
       }
       
@@ -865,10 +865,10 @@ int rsfwrite(ireal * a, IPNT rags, IPNT ran, char * fname, FILE * stream
       /* seek to write segment */
       /*      fprintf(stderr,"seek trace %d\n",i);*/
       if (!err && fseeko(fp,goffs[i]*sizeof(float) + cur_pos,SEEK_SET)) {
-	fprintf(stream,"Error: rsfwrite from fseeko at file offset %lld\n",(intmax_t)goffs[i]); 
+	fprintf(stream,"Error: rsfwrite from fseeko at file offset %lld\n",(long long int)goffs[i]); 
 	fprintf(stream,"possible cause: attempt to write off end of file\n");
 	fseeko(fp,0L,SEEK_END);
-	fprintf(stream,"file length = %lld:\n",(intmax_t)ftello(fp));
+	fprintf(stream,"file length = %lld:\n",(long long int)ftello(fp));
 	fprintf(stream,"note that new file can only be written in contiguous,\n");
 	fprintf(stream,"consecutive blocks\n");
 	err=E_FILE;
@@ -879,7 +879,7 @@ int rsfwrite(ireal * a, IPNT rags, IPNT ran, char * fname, FILE * stream
       }
       /* write out float buffer */
       if (!err && (na[0] != (int) fwrite(fbuf,sizeof(float),na[0],fp))) {
-	fprintf(stream,"Error: rsfwrite from fwrite at array offset %lld\n",(intmax_t)loffs[i]);
+	fprintf(stream,"Error: rsfwrite from fwrite at array offset %lld\n",(long long int)loffs[i]);
 	fprintf(stream,"failed to write %d words\n",na[0]);
 	err=E_FILE;
       }
@@ -907,11 +907,11 @@ int rsfwrite(ireal * a, IPNT rags, IPNT ran, char * fname, FILE * stream
 
       /* seek to write segment */
       if (!err && fseeko(fp,goffs[i]*sizeof(float) + cur_pos, SEEK_SET)) {
-	fprintf(stream,"Error: rsfwrite from fseeko at file offset %lld\n",(intmax_t)goffs[i]);
+	fprintf(stream,"Error: rsfwrite from fseeko at file offset %lld\n",(long long int)goffs[i]);
 	fprintf(stream,"possible cause: attempt to write off end of file\n");
 	j=0;
 	fseeko(fp,j,SEEK_END);
-	fprintf(stream,"file length = %lld:\n",(intmax_t)ftello(fp));
+	fprintf(stream,"file length = %lld:\n",(long long int)ftello(fp));
 	fprintf(stream,"note that new file can only be written in contiguous,\n");
 	fprintf(stream,"consecutive blocks\n");
 	err=E_FILE;
@@ -928,7 +928,7 @@ int rsfwrite(ireal * a, IPNT rags, IPNT ran, char * fname, FILE * stream
       }
       /* write out byte array */
       if (!err && (recsize_b != fwrite(buf,sizeof(char),recsize_b,fp))) {
-	fprintf(stream,"Error: rsfwrite from fwrite at array offset %lld\n",(intmax_t)loffs[i]);
+	fprintf(stream,"Error: rsfwrite from fwrite at array offset %lld\n",(long long int)loffs[i]);
 	err=E_FILE;
       }
